train contracting moved out of the base ensemble and into HIVE-COTE, to mirror contracting behaviour in HiveCote at least, still to be reviewed
